refactor: 메서드명 수정 및 패키지 이동
