+++ src/main/java/blackjack/domain/status/BlackjackStatus.java
@@ -1,7 +1,6 @@
 package blackjack.domain.status;
 
 import blackjack.domain.Score;
-import blackjack.domain.ParticipantScoreStatus;
 
 public enum BlackjackStatus {
     DEAD,
