+++ src/test/java/blackjack/domain/status/WinStatusTest.java
@@ -2,7 +2,6 @@ package blackjack.domain.status;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.ParticipantScoreStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
